llm model/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Image Processing
Pillow==10.1.0

# Deep Learning & ML
torch==2.1.1
transformers==4.35.2
optimum==1.16.1
optimum-onnxruntime==0.0.8

# Model Dependencies
onnxruntime==1.16.1
numpy>=1.24.0
requests>=2.31.0

# Optional but recommended
python-dotenv==1.0.0  # for environment variable management
pydantic==2.5.2  # for data validation
typing-extensions>=4.8.0  # for type hints 